Move Numerics to core
