+++ src/openrct2/core/Numerics.hpp
